Home File
This Wiki presence is nothing more than a kind of compendium of ideas and vision of a more complex fallout. I would really like to see all of these integrated into Horizon. It's not easy to describe something that doesn't exist in a way that everyone can understand. So don't hold it against me if every now and then whole nice grits are written. . Thanks to everyone who took on the bombing of ideas and possibly even made them a reality.